app/requirements.txt
fastapi
uvicorn
pymongo
celery
redis
pandas
python-dotenv
requests
pillow
